Removed empty tree.ts file & renamed common -> logic
