+++ src/main/javascript/modules/skill/woodcutting/logic.ts
